Adds more Macro fetching exceptions
If there are multiple code chunks present on a particular code chunks, then we would add a special CSS class (macro-code) to the actual Macro code.
So if there are multiple code-chunks having same CSS class as the main code, then we will select the main one first. If it's not present, then there is only one code block.